legalai-backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
PyJWT==2.9.0
passlib[argon2]==1.7.4
pgvector==0.2.5
celery==5.4.0
redis==5.0.8
requests==2.32.3
python-docx==1.1.2
reportlab==4.2.2
pypdf==4.3.1
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
gunicorn==22.0.0
flask-cors==4.0.1
openpyxl>=3.1.2
pillow>=10.3.0
pytesseract>=0.3.10
flask-swagger-ui
Redis
Flask-Limiter==3.8.0